reorganized code in app.js
