+++ public/javascripts/app.js
@@ -1,195 +1,10 @@
-import { handleErrors, addTaskInfoListeners, editListEventListener } from "./utils.js";
-import { fetchTask, fetchComments, postComment } from "./task-comments.js";
-
-// fetch user's tasks (all)
-
-export const fetchTasks = async () => {
-    const res = await fetch("/tasks")
-
-    if (res.status === 401) {
-        window.location.href = "/log-in";
-        return;
-      }
-
-    const { tasks, user } = await res.json();
-
-    const tasksListContainer = document.querySelector(".task-list");
-    const listName = `
-  <h2 class="task-list-header">All of <strong>${user.username}'s</strong> self-assigned tasks.</h2>
-  `
-    const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-      if (isCompleted === true) {
-        return `
-        <div class='task-info' id=${id}>
-            <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-            <label for=${id} id=${id} class="task-check-box">${description}</label>
-        </div>
-        `
-      } else {
-      return `<div class='task-info' id=${id}>
-                <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-                <label for=${id} id=${id} class="task-check-box">${description}</label>
-            </div>
-        `
-    }
-  })
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
-  await addTaskInfoListeners();
-  }
+import { handleErrors, addTaskInfoListeners } from "./utils.js";
+import { fetchTasks, fetchAssignTasks, fetchIncompleteTasks,fetchCompletedTasks } from "./fetch-tasks.js";
+import { search } from "./search.js";
+import { fetchContactTasks, addNewContact } from "./contacts.js";
+import { fetchLists } from "./lists.js";
 
 
-// fetch assigned tasks
-
-const fetchAssignTasks = async () => {
-  const res = await fetch("/tasks/assigned")
-
-  if (res.status === 401) {
-    window.location.href = "/log-in";
-    return;
-  }
-
-  const { tasks } = await res.json();
-
-  const assignedTaskContainer = document.querySelector(".assigned-list");
-  if(tasks) {
-    const listName = `
-    <h2 class="task-list-header">All tasks assigned to <strong></strong> by others.</h2>
-    `
-    const tasksHtml = tasks.map(({ id, description, User }) => `
-    <div class="assigned-grid">
-    <div class="task-info" id=${id}>
-        <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-        <label for=${id} id=${id} class="task-check-box">${description}</label>
-    </div>
-    <div>
-      <strong>Assigned by: ${User.username}</strong>
-    </div>
-    </div>
-    `)
-
-    assignedTaskContainer.innerHTML = listName + tasksHtml.join("")
-
-    await addTaskInfoListeners();
-  }
-
-}
-
-// fetch user's tasks (all)
-
-export const fetchIncompleteTasks = async () => {
-  const res = await fetch("/tasks/incomplete")
-
-  if (res.status === 401) {
-      window.location.href = "/log-in";
-      return;
-    }
-
-  const { tasks, user } = await res.json();
-
-  const tasksListContainer = document.querySelector(".task-list");
-  const listName = `
-  <h2 class="task-list-header"><strong>${user.username}'s</strong> incomplete tasks.</h2>
-  `
-  const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-    if (isCompleted === true) {
-      return `
-      <div class='task-info' id=${id}>
-          <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-          <label for=${id} id=${id} class="task-check-box">${description}</label>
-      </div>
-      `
-    } else {
-    return `<div class='task-info' id=${id}>
-              <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-              <label for=${id} id=${id} class="task-check-box">${description}</label>
-          </div>
-      `
-  }
-})
-
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
-
-  await addTaskInfoListeners();
-}
-
-// fetch user's completed tasks
-
-const fetchCompletedTasks = async () => {
-  const res = await fetch("/tasks/complete")
-
-  if (res.status === 401) {
-      window.location.href = "/log-in";
-      return;
-    }
-
-  const { tasks, user } = await res.json();
-
-  const tasksListContainer = document.querySelector(".task-list");
-  const listName = `
-  <h2 class="task-list-header"><strong>${user.username}'s</strong> completed tasks.</h2>
-  `
-  const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-    if (isCompleted === true) {
-      return `
-      <div class='task-info' id=${id}>
-          <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-          <label for=${id} id=${id} class="task-check-box">${description}</label>
-      </div>
-      `
-    } else {
-    return `<div class='task-info' id=${id}>
-              <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-              <label for=${id} id=${id} class="task-check-box">${description}</label>
-          </div>
-      `
-  }
-})
-
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
-
-  await addTaskInfoListeners();
-}
-
-// fetch user's search query
-
-const search = async (searchValue) => {
-
-    const res = await fetch(`/search/${searchValue}`, {
-      method: "GET"
-    })
-
-    const { results } = await res.json();
-    if (res.status === 401) {
-      window.location.href = "/log-in";
-      return;
-    }
-
-
-  const tasksListContainer = document.querySelector(".task-list");
-  const listName = `
-  <h2 class="task-list-header">Search Results</h2>
-  `
-  const tasksHtml = results.map(({ id, description, isCompleted, Lists }) => {
-    if (isCompleted === true) {
-      return `
-      <div class='task-info' id=${id}>
-          <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-          <label for=${id} id=${id} class="task-check-box">${description}</label>
-      </div>
-      `
-    } else {
-    return `<div class='task-info' id=${id}>
-              <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-              <label for=${id} id=${id} class="task-check-box">${description}</label>
-          </div>
-      `
-  }
-})
-
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
-
-  await addTaskInfoListeners();
-}
 
 
 // // Change the color of incomplete/complete tabs when fetching their respective lists
@@ -201,6 +16,8 @@ const search = async (searchValue) => {
 
 // toggle between incomplete and completed tasks
 // incomplete button
+
+
 const incompleteTaskList = document.querySelector('#incomplete')
 incompleteTaskList.addEventListener("click", async (e) => {
   await fetchIncompleteTasks();
@@ -224,80 +41,8 @@ completeTaskList.addEventListener("click", async (e) => {
   clearAssignedList.innerHTML = ``;
 })
 
-// shows tasks that user assigns to their contacts
 
-const fetchContactTasks = async (id) => {
-  const res = await fetch(`/tasks/task/${id.id}`)
-  if (res.status === 401) {
-      window.location.href = "/log-in";
-      return;
-    }
 
-  const { tasks, user, isContact } = await res.json();
-  const tasksListContainer = document.querySelector(".task-list");
-  let listName = ``;
-  if(!isContact) {
-    listName = `
-    <h2 class="task-list-header">All of <strong>${user.username}'s</strong> self-assigned tasks.</h2>
-    `
-    //await fetchAssignTasks();
-  } else {
-    listName = `
-  <h2 class="task-list-header">Tasks that you've assigned to <strong>${user.username}</strong>.</h2>
-  `
-  const clearAssignedList = document.querySelector('.assigned-list')
-  clearAssignedList.innerHTML = ``;
-  }
-  const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-    if (isCompleted === true) {
-      return `
-      <div class='task-info' id=${id}>
-          <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-          <label for=${id} id=${id} class="task-check-box">${description}</label>
-      </div>
-      `
-    } else {
-    return `<div class='task-info' id=${id}>
-              <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-              <label for=${id} id=${id} class="task-check-box">${description}</label>
-          </div>
-      `
-  }
-})
-
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
-
-  await addTaskInfoListeners();
-
-}
-
-//dynamically add new contact to the sidebar
-const addNewContact = async (id) => {
-  const res = await fetch(`/users/${id}`, {
-    method: "GET"
-  })
-
-  if (res.status === 401) {
-    window.location.href = "/log-in";
-    return;
-  }
-
-  const { userInfo } = await res.json();
-  const addNewContact = `
-  <div class="list-grid">
-    <div>
-        <li class="contact-list" id=${userInfo.id}>${userInfo.username}</li>
-    </div>
-    <div>
-        <a class="delete-contact" id=${userInfo.id}>-</a>
-    </div>
-  </div>
-  `
-  const contactContainer = document.querySelector('.contact-list-sidebar');
-  const node = document.createElement("div")
-  node.innerHTML = addNewContact;
-  contactContainer.appendChild(node);
-}
 
 document.addEventListener("DOMContentLoaded", async () => {
     try {
@@ -482,76 +227,7 @@ deleteContact.addEventListener("click", async (e) => {
 
 })
 
-export const fetchLists = async () => {
-  const res = await fetch('/lists/')
 
-  if (res.status === 401) {
-      window.location.href = "/log-in";
-      return;
-    }
-
-  const { allLists } = await res.json();
-
-  const listContainer = document.querySelector(".lists-grid-container");
-  const listHtml = allLists.map(({ id, title }) => `
-  <div id="delete-hover-parent" class='list-grid'>
-  <div class="list-info">
-    <li class='list-lists' id=${id}>${title}</li>
-  </div>
-  <div > <a class='delete-list' id=${id}> - </a> </div>
-  </div>
-  `)
-
-  listContainer.innerHTML = listHtml.join("");
-
-  const listLists = document.querySelectorAll('.list-lists');
-
-    listLists.forEach((list) => {
-      list.addEventListener('click', async(e) => {
-        e.stopPropagation();
-        const listId = list.id;
-
-        const res = await fetch(`/lists/${listId}/tasks`);
-
-        const { tasks } = await res.json();
-
-        const tasksContainer = document.querySelector('.task-list');
-
-        const listTitle = `
-        <div class="list-title" id="${listId}">
-          <h2 class="task-list-header">${list.innerText}</h2>
-          <button class="edit-list-button edit-button-modal button-modal" id="${listId}">Edit List</button>
-        </div>
-        `
-
-        const tasksHtml = tasks.map((task) => {
-          if (task.Task.isCompleted === true) {
-            return `
-              <div class="task-info" id=${task.Task.id}>
-                <input type="checkbox" class="task-check-box" id=${task.Task.id} name=${task.Task.id} checked>
-                <label for=${task.Task.id} id=${task.Task.id} class="task-check-box">${task.Task.description}</label>
-              </div>
-            `
-          } else {
-
-            return `
-              <div class="task-info" id=${task.Task.id}>
-                <input type="checkbox" class="task-check-box" id=${task.Task.id} name=${task.Task.id}>
-                <label for=${task.Task.id} id=${task.Task.id} class="task-check-box">${task.Task.description}</label>
-              </div>
-          `
-          }
-        })
-
-        tasksContainer.innerHTML = listTitle + tasksHtml.join('');
-
-
-
-        await editListEventListener();
-        await addTaskInfoListeners();
-      })
-    })
-}
 
 
 // delete a list
