+++ public/javascripts/utils.js
@@ -1,5 +1,8 @@
-import { fetchTask, fetchComments, postComment } from "./task-comments.js";
-import { fetchLists } from "./app.js"
+import { fetchTask } from "./task-info-container.js";
+import { fetchLists } from "./lists.js"
+import { fetchComments, postComment } from "./comments.js";
+
+
 
 export const handleErrors = async (err) => {
     if (err.status >= 400 && err.status < 600) {
@@ -29,68 +32,6 @@ export const handleErrors = async (err) => {
       );
     }
   };
-
-export const editListEventListener = async () => {
-  const editListTitle = document.querySelector(".edit-list-button");
-        editListTitle.addEventListener('click', async(e) => {
-          let listId = e.target.id;
-
-          const listToUpdate = await fetch(`/lists/${listId}`);
-
-          const { listName } = await listToUpdate.json();
-
-          const listForm = document.querySelector('.updateList');
-          listForm.innerHTML = `
-          <div class="cloud"></div>
-          <div class="edit-list-pop">
-            <h2 class="modal-header">Edit List Name</h2>
-            <div id='list-edit'>
-              <form class='list-edit-form modal-form'>
-              <input type='text' class='list-edit modal-input' id='title' name='title' placeholder=${listName.title}>
-              <label for='title' class='list-label'${listName.title} </label>
-              <div>
-              <button class='submitButton button-modal' id='${listId}'>Submit</button>
-
-              <button class='editCancelButton button-modal' id='${listId}'>Cancel</button>
-              </div>
-              </form>
-            </div>
-            </div>
-          `
-
-          const listUpdate = document.querySelector('.list-edit-form')
-          listUpdate.addEventListener('submit', async (e) => {
-            e.preventDefault();
-            e.stopPropagation();
-            const formData = new FormData(listUpdate);
-            const title = formData.get('title')
-            const body = { title }
-            const updatedList = await fetch(`/lists/${listId}`, {
-              method: 'PATCH',
-              body: JSON.stringify(body),
-              headers: {
-                'Content-Type': 'application/json'
-              }
-            })
-
-            const { list } = await updatedList.json();
-
-            const listHeader = document.querySelector('.task-list-header');
-            listHeader.innerText = list.title;
-
-            listForm.innerHTML = '';
-            await fetchLists();
-          })
-
-
-        const cancelButton = document.querySelector('.editCancelButton');
-        cancelButton.addEventListener('click', (e) => {
-          e.preventDefault();
-          e.stopPropagation();
-          listForm.innerHTML = '';
-        })
-      })
-}
 
 
 
@@ -211,4 +152,68 @@ export const addTaskInfoListeners = async () => {
   })
 
 
+}
+
+
+
+export const editListEventListener = async () => {
+  const editListTitle = document.querySelector(".edit-list-button");
+        editListTitle.addEventListener('click', async(e) => {
+          let listId = e.target.id;
+
+          const listToUpdate = await fetch(`/lists/${listId}`);
+
+          const { listName } = await listToUpdate.json();
+
+          const listForm = document.querySelector('.updateList');
+          listForm.innerHTML = `
+          <div class="cloud"></div>
+          <div class="edit-list-pop">
+            <h2 class="modal-header">Edit List Name</h2>
+            <div id='list-edit'>
+              <form class='list-edit-form modal-form'>
+              <input type='text' class='list-edit modal-input' id='title' name='title' placeholder=${listName.title}>
+              <label for='title' class='list-label'${listName.title} </label>
+              <div>
+              <button class='submitButton button-modal' id='${listId}'>Submit</button>
+
+              <button class='editCancelButton button-modal' id='${listId}'>Cancel</button>
+              </div>
+              </form>
+            </div>
+            </div>
+          `
+
+          const listUpdate = document.querySelector('.list-edit-form')
+          listUpdate.addEventListener('submit', async (e) => {
+            e.preventDefault();
+            e.stopPropagation();
+            const formData = new FormData(listUpdate);
+            const title = formData.get('title')
+            const body = { title }
+            const updatedList = await fetch(`/lists/${listId}`, {
+              method: 'PATCH',
+              body: JSON.stringify(body),
+              headers: {
+                'Content-Type': 'application/json'
+              }
+            })
+
+            const { list } = await updatedList.json();
+
+            const listHeader = document.querySelector('.task-list-header');
+            listHeader.innerText = list.title;
+
+            listForm.innerHTML = '';
+            await fetchLists();
+          })
+
+
+        const cancelButton = document.querySelector('.editCancelButton');
+        cancelButton.addEventListener('click', (e) => {
+          e.preventDefault();
+          e.stopPropagation();
+          listForm.innerHTML = '';
+        })
+      })
 }
